refactor: page styles cleanup
