+++ src/pages/add-order-page/add-order-page.tsx
@@ -2,7 +2,7 @@ import { Box } from '@mui/material';
 
 import { Header } from '@/layouts/header';
 
-import { useStyles } from './styles';
+import { useStyles } from '../styles';
 
 /* eslint-disable import/no-default-export */
 export default function AddOrderPage(): React.ReactNode {
