+++ src/pages/home-page/home-page.tsx
@@ -2,7 +2,7 @@ import { Box } from '@mui/material';
 
 import { Header } from '@/layouts/header';
 
-import { useStyles } from './styles';
+import { useStyles } from '../styles';
 
 /* eslint-disable import/no-default-export */
 export default function HomePage(): React.ReactNode {
