--- src/pages/material-page/styles.ts
@@ -1,15 +1,0 @@
-import { tss } from '@/styles/theme';
-
-export const useStyles = tss.create(({ theme }) => {
-  return {
-    wrapper: {
-      color: theme.palette.main.primary,
-      display: 'flex',
-      flexDirection: 'column',
-      alignItems: 'flex-start',
-      backgroundColor: theme.palette.light.secondary,
-      marginTop: '12px',
-      borderTopLeftRadius: '28px',
-    },
-  };
-});
