+++ src/pages/order-page/order-page.tsx
@@ -6,9 +6,9 @@ import { TabsSection } from '@/shared/tabs-section';
 
 import { Header } from '@/layouts/header';
 
-import { ORDER_TABS } from './constants';
+import { useStyles } from '../styles';
 
-import { useStyles } from './styles';
+import { ORDER_TABS } from './constants';
 
 /* eslint-disable import/no-default-export */
 export default function OrderPage(): React.ReactNode {
